Create endpoints for product categories, best products and personal recommendations

+++ database/migrations/2023_09_18_204547_create_products_table.php
@@ -26,7 +26,7 @@ return new class extends Migration {
             $table->boolean('draft')->default(false);
             $table->integer('quantity')->default(0);
             $table->tinyInteger('currency')->default(CurrencyEnum::USD->value);
-            $table->integer('cost');
+            $table->float('cost');
             $table->timestamps();
         });
     }
